fastapi_solution/requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.9.1
fastapi==0.61.1
orjson==3.8.5
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.17.0
gunicorn==20.1.0
python-dotenv==0.21.0